change float to double

+++ inc/draw_ellipse.h
@@ -7,28 +7,28 @@
 #include <string>
 void sdl_ellipse(SDL_Renderer* r, int x0, int y0, int radiusX, int radiusY)
 {
-    float pi  = 3.14159265358979323846264338327950288419716939937510;
-    float pih = pi / 2.0; //half of pi
+    double pi  = 3.14159265358979323846264338327950288419716939937510;
+    double pih = pi / 2.0; //half of pi
 
     //drew  28 lines with   4x4  circle with precision of 150 0ms
     //drew 132 lines with  25x14 circle with precision of 150 0ms
     //drew 152 lines with 100x50 circle with precision of 150 3ms
     const int prec = 27; // precision value; value of 1 will draw a diamond, 27 makes pretty smooth circles.
-    float theta = 0;     // angle that will be increased each loop
+    double theta = 0;     // angle that will be increased each loop
 
     //starting point
-    int x  = (float)radiusX * cos(theta);//start point
-    int y  = (float)radiusY * sin(theta);//start point
+    int x  = (double)radiusX * cos(theta);//start point
+    int y  = (double)radiusY * sin(theta);//start point
     int x1 = x;
     int y1 = y;
 
     //repeat until theta >= 90;
-    float step = pih/(float)prec; // amount to add to theta each time (degrees)
+    double step = pih/(double)prec; // amount to add to theta each time (degrees)
     for(theta=step;  theta <= pih;  theta+=step)//step through only a 90 arc (1 quadrant)
     {
         //get new point location
-        x1 = (float)radiusX * cosf(theta) + 0.5; //new point (+.5 is a quick rounding method)
-        y1 = (float)radiusY * sinf(theta) + 0.5; //new point (+.5 is a quick rounding method)
+        x1 = (double)radiusX * cosf(theta) + 0.5; //new point (+.5 is a quick rounding method)
+        y1 = (double)radiusY * sinf(theta) + 0.5; //new point (+.5 is a quick rounding method)
 
         //draw line from previous point to new point, ONLY if point incremented
         if( (x != x1) || (y != y1) )//only draw if coordinate changed
